Modified : Functionality Modified to Find Cheapest Hotel for Regular Customer based on Weekdays and Weekend Rates

+++ Hotel.java
@@ -4,13 +4,13 @@ public class Hotel {
 
 	private String hotelName;
 	private int weekdaysRatesRegularCustomer, weekendRatesRegularCustomer;
-	public long TotalRate;
+	public int TotalRate;
 
-	public long getTotalRate() {
+	public int getTotalRate() {
 		return TotalRate;
 	}
 
-	public void setTotalRate(long totalRate) {
+	public void setTotalRate(int totalRate) {
 		TotalRate = totalRate;
 	}
